mv time series, tidy up
