+++ weldx/core.py
@@ -1,474 +1,7 @@
 """Collection of common classes and functions."""
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any, Tuple, Union
-from warnings import warn
-
-import numpy as np
-import pandas as pd
-import pint
-import xarray as xr
-
-import weldx.util as ut
-from weldx.constants import Q_, UNITS_KEY
-from weldx.core.math_expression import MathematicalExpression
-from weldx.time import Time, TimeDependent, types_time_like, types_timestamp_like
-from weldx.util import check_matplotlib_available
-
-if TYPE_CHECKING:  # pragma: no cover
-    import matplotlib.axes
-
-    from weldx.types import UnitLike
-
-__all__ = ["TimeSeries"]
-
-_me_parameter_types = Union[pint.Quantity, str, Tuple[pint.Quantity, str], xr.DataArray]
-
-
 # TimeSeries ---------------------------------------------------------------------------
-
-
-class TimeSeries(TimeDependent):
-    """Describes the behaviour of a quantity in time."""
-
-    _valid_interpolations = [
-        "step",
-        "linear",
-        "nearest",
-        "zero",
-        "slinear",
-        "quadratic",
-        "cubic",
-    ]
-
-    def __init__(
-        self,
-        data: Union[pint.Quantity, MathematicalExpression],
-        time: types_time_like = None,
-        interpolation: str = None,
-        reference_time: types_timestamp_like = None,
-    ):
-        """Construct a TimSeries.
-
-        Parameters
-        ----------
-        data:
-            Either a pint.Quantity or a weldx.MathematicalExpression. If a mathematical
-            expression is chosen, it is only allowed to have a single free variable,
-            which represents time.
-        time:
-            An instance of pandas.TimedeltaIndex if a quantity is passed and 'None'
-            otherwise.
-        interpolation:
-            A string defining the desired interpolation method. This is only relevant if
-            a quantity is passed as data. Currently supported interpolation methods are:
-            'step', 'linear'.
-
-        """
-        self._data: Union[MathematicalExpression, xr.DataArray] = None
-        self._time_var_name = None  # type: str
-        self._shape = None
-        self._units = None
-        self._interp_counter = 0
-        self._reference_time = None
-
-        if isinstance(data, (pint.Quantity, xr.DataArray)):
-            self._initialize_discrete(data, time, interpolation, reference_time)
-        elif isinstance(data, MathematicalExpression):
-            self._init_expression(data, reference_time)
-        else:
-            raise TypeError(f'The data type "{type(data)}" is not supported.')
-
-    def __eq__(self, other: Any) -> bool:
-        """Return the result of a structural equality comparison with another object.
-
-        If the other object is not a 'TimeSeries' this function always returns 'False'.
-
-        Parameters
-        ----------
-        other:
-            Other object.
-
-        Returns
-        -------
-        bool:
-           'True' if the compared object is also a 'TimeSeries' and equal to
-            this instance, 'False' otherwise
-
-        """
-        if not isinstance(other, TimeSeries):
-            return False
-        if not isinstance(self.data, MathematicalExpression):
-            if not isinstance(other.data, pint.Quantity):
-                return False
-            return self._data.identical(other.data_array)  # type: ignore
-
-        return self._data == other.data
-
-    def __repr__(self):
-        """Give __repr__ output."""
-        representation = "<TimeSeries>"
-        if isinstance(self._data, xr.DataArray):
-            if self._data.shape[0] == 1:
-                representation += f"\nConstant value:\n\t{self.data.magnitude[0]}\n"
-            else:
-                representation += (
-                    f"\nTime:\n\t{self.time}\n"
-                    f"Values:\n\t{self.data.magnitude}\n"
-                    f'Interpolation:\n\t{self._data.attrs["interpolation"]}\n'
-                )
-        else:
-            representation += self.data.__repr__().replace(
-                "<MathematicalExpression>", ""
-            )
-        return representation + f"Units:\n\t{self.units}\n"
-
-    @staticmethod
-    def _check_data_array(data_array: xr.DataArray):
-        """Raise an exception if the 'DataArray' can't be used as 'self._data'."""
-        try:
-            ut.xr_check_coords(data_array, dict(time={"dtype": ["timedelta64[ns]"]}))
-        except (KeyError, TypeError, ValueError) as e:
-            raise type(e)(
-                "The provided 'DataArray' does not match the required pattern. It "
-                "needs to have a dimension called 'time' with coordinates of type "
-                "'timedelta64[ns]'. The error reported by the comparison function was:"
-                f"\n{e}"
-            )
-
-        if not isinstance(data_array.data, pint.Quantity):
-            raise TypeError("The data of the 'DataArray' must be a 'pint.Quantity'.")
-
-    @staticmethod
-    def _create_data_array(
-        data: Union[pint.Quantity, xr.DataArray], time: Time
-    ) -> xr.DataArray:
-        return (
-            xr.DataArray(data=data)
-            .rename({"dim_0": "time"})
-            .assign_coords({"time": time.as_data_array()})
-        )
-
-    def _initialize_discrete(
-        self,
-        data: Union[pint.Quantity, xr.DataArray],
-        time: types_time_like = None,
-        interpolation: str = None,
-        reference_time=None,
-    ):
-        """Initialize the internal data with discrete values."""
-        # set default interpolation
-        if interpolation is None:
-            interpolation = "step"
-
-        if isinstance(data, xr.DataArray):
-            self._check_data_array(data)
-            data = data.transpose("time", ...)
-            self._data = data
-        else:
-            # expand dim for scalar input
-            data = Q_(data)
-            if not np.iterable(data):
-                data = np.expand_dims(data, 0)
-
-            # constant value case
-            if time is None:
-                time = pd.Timedelta(0)
-            time = Time(time, reference_time)
-
-            self._data = self._create_data_array(data, time)
-        self.interpolation = interpolation
-
-    def _init_expression(self, data, reference_time):
-        """Initialize the internal data with a mathematical expression."""
-        if data.num_variables != 1:
-            raise Exception(
-                "The mathematical expression must have exactly 1 free "
-                "variable that represents time."
-            )
-
-        # check that the expression can be evaluated with a time quantity
-        time_var_name = data.get_variable_names()[0]
-        try:
-            eval_data = data.evaluate(**{time_var_name: Q_(1, "second")}).data
-            self._units = eval_data.units
-            if np.iterable(eval_data):
-                self._shape = eval_data.shape
-            else:
-                self._shape = (1,)
-        except pint.errors.DimensionalityError:
-            raise Exception(
-                "Expression can not be evaluated with "
-                '"weldx.Quantity(1, "seconds")"'
-                ". Ensure that every parameter posses the correct unit."
-            )
-
-        # assign internal variables
-        self._data = data
-        self._time_var_name = time_var_name
-        if reference_time is not None:
-            self._reference_time = pd.Timestamp(reference_time)
-
-        # check that all parameters of the expression support time arrays
-        try:
-            self.interp_time(Q_([1, 2], "second"))
-            self.interp_time(Q_([1, 2, 3], "second"))
-        except Exception as e:
-            raise Exception(
-                "The expression can not be evaluated with arrays of time deltas. "
-                "Ensure that all parameters that are multiplied with the time "
-                "variable have an outer dimension of size 1. This dimension is "
-                "broadcasted during multiplication. The original error message was:"
-                f' "{str(e)}"'
-            )
-
-    def _interp_time_discrete(self, time: Time) -> xr.DataArray:
-        """Interpolate the time series if its data is composed of discrete values."""
-        data = self._data
-        if self.time is None and time.is_absolute:
-            data = data.weldx.reset_reference_time(time.reference_time)  # type: ignore
-
-        return ut.xr_interp_like(
-            data,
-            time.as_data_array(),
-            method=self.interpolation,
-            assume_sorted=False,
-            broadcast_missing=False,
-        )
-
-    def _interp_time_expression(self, time: Time, time_unit: str) -> xr.DataArray:
-        """Interpolate the time series if its data is a mathematical expression."""
-        time_q = time.as_quantity(unit=time_unit)
-        if len(time_q.shape) == 0:
-            time_q = np.expand_dims(time_q, 0)  # type: ignore[assignment]
-
-        time_xr = xr.DataArray(time_q, dims=["time"])
-
-        # evaluate expression
-        data = self._data.evaluate(**{self._time_var_name: time_xr})
-        return data.assign_coords({"time": time.as_data_array()})
-
-    @property
-    def data(self) -> Union[pint.Quantity, MathematicalExpression]:
-        """Return the data of the TimeSeries.
-
-        This is either a set of discrete values/quantities or a mathematical expression.
-
-        Returns
-        -------
-        pint.Quantity:
-            Underlying data array.
-        MathematicalExpression:
-            A mathematical expression describing the time dependency
-
-        """
-        if isinstance(self._data, xr.DataArray):
-            return self._data.data
-        return self._data
-
-    @property
-    def data_array(self) -> Union[xr.DataArray, None]:
-        """Return the internal data as 'xarray.DataArray'.
-
-        If the TimeSeries contains an expression, 'None' is returned.
-
-        Returns
-        -------
-        xarray.DataArray:
-            The internal data as 'xarray.DataArray'
-
-        """
-        if isinstance(self._data, xr.DataArray):
-            return self._data
-        return None
-
-    @property
-    def interpolation(self) -> Union[str, None]:
-        """Return the interpolation.
-
-        Returns
-        -------
-        str:
-            Interpolation of the TimeSeries
-
-        """
-        if isinstance(self._data, xr.DataArray):
-            return self._data.attrs["interpolation"]
-        return None
-
-    @interpolation.setter
-    def interpolation(self, interpolation):
-        if isinstance(self._data, xr.DataArray):
-            if interpolation not in self._valid_interpolations:
-                raise ValueError(
-                    "A valid interpolation method must be specified if discrete "
-                    f'values are used. "{interpolation}" is not supported'
-                )
-            if self.time is None and interpolation != "step":
-                interpolation = "step"
-            self.data_array.attrs["interpolation"] = interpolation
-
-    @property
-    def is_discrete(self) -> bool:
-        """Return `True` if the time series is described by discrete values."""
-        return not self.is_expression
-
-    @property
-    def is_expression(self) -> bool:
-        """Return `True` if the time series is described by an expression."""
-        return isinstance(self.data, MathematicalExpression)
-
-    @property
-    def time(self) -> Union[None, Time]:
-        """Return the data's timestamps.
-
-        Returns
-        -------
-        pandas.TimedeltaIndex:
-            Timestamps of the  data
-
-        """
-        if isinstance(self._data, xr.DataArray) and len(self._data.time) > 1:
-            return Time(self._data.time.data, self.reference_time)
-        return None
-
-    @property
-    def reference_time(self) -> Union[pd.Timestamp, None]:
-        """Get the reference time."""
-        if self.is_discrete:
-            return self._data.weldx.time_ref  # type: ignore[union-attr]
-        return self._reference_time
-
-    def interp_time(
-        self, time: Union[pd.TimedeltaIndex, pint.Quantity, Time], time_unit: str = "s"
-    ) -> TimeSeries:
-        """Interpolate the TimeSeries in time.
-
-        If the internal data consists of discrete values, an interpolation with the
-        prescribed interpolation method is performed. In case of mathematical
-        expression, the expression is evaluated for the given timestamps.
-
-        Parameters
-        ----------
-        time:
-            The time values to be used for interpolation.
-        time_unit:
-            Only important if the time series is described by an expression and a
-            'pandas.TimedeltaIndex' is passed to this function. In this case, time is
-            converted to a quantity with the provided unit. Even though pint handles
-            unit prefixes automatically, the accuracy of the results can be heavily
-            influenced if the provided unit results in extreme large or
-            small values when compared to the parameters of the expression.
-
-        Returns
-        -------
-        TimeSeries :
-            A new `TimeSeries` object containing the interpolated data.
-
-        """
-        if self._interp_counter > 0:
-            warn(
-                "The data of the time series has already been interpolated "
-                f"{self._interp_counter} time(s)."
-            )
-
-        # prepare timedelta values for internal interpolation
-        time = Time(time)
-        time_interp = Time(time, self.reference_time)
-
-        if isinstance(self._data, xr.DataArray):
-            dax = self._interp_time_discrete(time_interp)
-            ts = TimeSeries(data=dax.data, time=time, interpolation=self.interpolation)
-        else:
-            dax = self._interp_time_expression(time_interp, time_unit)
-            ts = TimeSeries(data=dax, interpolation=self.interpolation)
-
-        ts._interp_counter = self._interp_counter + 1
-        return ts
-
-    @check_matplotlib_available
-    def plot(
-        self,
-        time: Union[pd.TimedeltaIndex, pint.Quantity] = None,
-        axes: "matplotlib.axes.Axes" = None,  # noqa: F821
-        data_name: str = "values",
-        time_unit: UnitLike = None,
-        **mpl_kwargs,
-    ) -> "matplotlib.axes.Axes":  # noqa: F821
-        """Plot the `TimeSeries`.
-
-        Parameters
-        ----------
-        time :
-            The points in time that should be plotted. This is an optional parameter for
-            discrete `TimeSeries` but mandatory for expression based TimeSeries.
-        axes :
-            An optional matplotlib axes object
-        data_name :
-            Name of the data that will appear in the y-axis label
-        mpl_kwargs :
-            Key word arguments that are passed to the matplotlib plot function
-        time_unit :
-            The desired time unit for the plot. If `None` is provided, the internally
-            stored unit will be used.
-
-        Returns
-        -------
-         matplotlib.axes.Axes :
-            The matplotlib axes object that was used for the plot
-
-        """
-        import matplotlib.pyplot as plt
-
-        if axes is None:
-            _, axes = plt.subplots()
-        if self.is_expression or time is not None:
-            return self.interp_time(time).plot(
-                axes=axes, data_name=data_name, time_unit=time_unit, **mpl_kwargs
-            )
-
-        time = Time(self.time, self.reference_time).as_quantity()
-        if time_unit is not None:
-            time = time.to(time_unit)
-
-        axes.plot(time.m, self._data.data.m, **mpl_kwargs)  # type: ignore
-        axes.set_xlabel(f"t in {time.u:~}")
-        y_unit_label = ""
-        if self.units not in ["", "dimensionless"]:
-            y_unit_label = f" in {self.units:~}"
-        axes.set_ylabel(data_name + y_unit_label)
-
-        return axes
-
-    @property
-    def shape(self) -> tuple:
-        """Return the shape of the TimeSeries data.
-
-        For mathematical expressions, the shape does not contain the time axis.
-
-        Returns
-        -------
-        Tuple:
-            Tuple describing the data's shape
-
-        """
-        if isinstance(self._data, xr.DataArray):
-            return self._data.shape
-        return self._shape
-
-    @property
-    def units(self) -> pint.Unit:
-        """Return the units of the TimeSeries Data.
-
-        Returns
-        -------
-        pint.Unit:
-            The unit of the `TimeSeries`
-
-        """
-        if isinstance(self._data, xr.DataArray):
-            return self._data.data.units
-        return self._units
 
 
 # GenericSeries ------------------------------------------------------------------------
@@ -477,19 +10,3 @@ class TimeSeries(TimeDependent):
 #  - pandas time types in TimeSeries vs GenericSeries
 #
 #  - add doctests (examples)
-
-
-def _quantity_to_coord_tuple(
-    v: pint.Quantity, dim
-) -> tuple[str, np.ndarray, dict[str, pint.Unit]]:
-    return (dim, v.m, {UNITS_KEY: v.u})
-
-
-def _quantity_to_xarray(v: pint.Quantity, dim: str = None) -> xr.DataArray:
-    """Convert a single quantity into a formatted xarray dataarray."""
-    return xr.DataArray(v, dims=dim)
-
-
-def _quantities_to_xarray(q_dict: dict[str, pint.Quantity]) -> dict[str, xr.DataArray]:
-    """Convert a str:Quantity mapping into a mapping of `xarray.DataArray`."""
-    return {k: _quantity_to_xarray(v, k) for k, v in q_dict.items()}
