+++ weldx/core/__init__.py
@@ -1,3 +1,5 @@
+"""Collection of common classes and functions."""
+
 from .generic_series import GenericSeries
 from .math_expression import MathematicalExpression
 from .spatial_series import SpatialSeries
